requirements.txt
numpy==2.4.3
pandas==3.0.1
matplotlib==3.10.8
seaborn==0.13.2
yfinance==1.2.0
fpdf2==2.8.7
